+++ src/analyze/src/components/dashboards/Dashboards.vue
@@ -12,30 +12,28 @@
             <a class='button is-secondary is-fullwidth'
                 @click="setAddDashboard(true)">Create Dashboard</a>
           </div>
-
-          <div v-for="dashboard in dashboards"
-              class='panel-block'
-              :class="{'is-active': isActive(dashboard)}"
-              :key="dashboard.id"
-              @click="getDashboard(dashboard)">
-            <div>
-              <div>{{dashboard.name}}</div>
-              <div v-if="dashboard.id === activeDashboard.id">
-                <small>Reports ({{activeDashboard.reportIds.length}})</small>
-                <ul>
-                  <li v-for="report in reports" :key="report.id">
-                    <label for="'checkbox-' + report.id"
-                            @click="toggleReportInDashboard(report)">
-                      <input type="checkbox"
-                            :id="'checkbox-' + report.id"
-                            :checked="isReportInActiveDashboard(report)">
-                      {{report.name}}</label>
-                  </li>
-                </ul>
+          <div v-if="hasDashboards">
+            <div v-for="dashboard in dashboards"
+                class='panel-block'
+                :class="{'is-active': isActive(dashboard)}"
+                :key="dashboard.id"
+                @click="getDashboard(dashboard)">
+              <div>
+                <div>{{dashboard.name}}</div>
+                <div v-if="getIsActiveDashboardMatch(dashboard)">
+                  <small>Reports ({{activeDashboard.reportIds.length}})</small>
+                  <ul>
+                    <li v-for="report in reports" :key="report.id">
+                      <label @click="toggleReportInDashboard(report)">
+                        <input type="checkbox"
+                              :checked="isReportInActiveDashboard(report)">
+                        {{report.name}}</label>
+                    </li>
+                  </ul>
+                </div>
               </div>
             </div>
           </div>
-
         </nav>
 
         <div class="column is-three-quarters">
@@ -82,14 +80,13 @@
 </template>
 
 <script>
-import { mapState, mapActions } from 'vuex';
+import { mapState, mapGetters, mapActions } from 'vuex';
 import Chart from '../designs/Chart';
 
 export default {
   name: 'Dashboards',
   created() {
-    this.getDashboards();
-    this.getReports();
+    this.initialize(this.$route.params.slug);
   },
   components: {
     Chart,
@@ -103,13 +100,15 @@ export default {
       'reports',
       'saveDashboardSettings',
     ]),
+    ...mapGetters('dashboards', [
+      'hasDashboards',
+      'getIsActiveDashboardMatch',
+    ]),
   },
   methods: {
     ...mapActions('dashboards', [
-      'getDashboards',
+      'initialize',
       'getDashboard',
-      'getReports',
-      'getActiveDashboardReportsWithQueryResults',
       'setAddDashboard',
     ]),
     isActive(dashboard) {
@@ -129,11 +128,6 @@ export default {
         reportId: report.id,
         dashboardId: this.activeDashboard.id,
       });
-    },
-  },
-  watch: {
-    activeDashboard() {
-      this.getActiveDashboardReportsWithQueryResults();
     },
   },
 };