+++ src/analyze/src/router/index.js
@@ -4,7 +4,9 @@ import NewProjectForm from '@/components/projects/NewProjectForm';
 import Repo from '@/components/repos/Repo';
 import Design from '@/components/designs/Design';
 import Dashboards from '@/components/dashboards/Dashboards';
-import Settings from '@/components/settings/Settings';
+import Settings from '@/components/settings/Main';
+import SettingsDatabase from '@/components/settings/Database';
+import SettingsRoles from '@/components/settings/Roles';
 import Extract from '@/components/extract/Extract';
 import Load from '@/components/load/Load';
 import Transform from '@/components/transform/Transform';
@@ -50,8 +52,13 @@ export default new Router({
       component: Repo,
     },
     {
-      path: '/design/:model/:design',
+      path: '/analyze/:model/:design',
       name: '',
+      component: Design,
+    },
+    {
+      path: '/analyze/:model/:design/reports/report/:slug',
+      name: 'Report',
       component: Design,
     },
     {
@@ -60,14 +67,21 @@ export default new Router({
       component: Dashboards,
     },
     {
-      path: '/dashboards/:id',
-      name: 'Dashboards',
+      path: '/dashboards/dashboard/:slug',
+      name: 'Dashboard',
       component: Dashboards,
     },
     {
       path: '/settings',
-      name: '',
+      name: 'Settings',
       component: Settings,
+      children: [{
+        path: 'roles',
+        component: SettingsRoles,
+      }, {
+        path: 'database',
+        component: SettingsDatabase,
+      }],
     },
     {
       path: '/orchestrations',