+++ src/meltano/api/controllers/dashboards_helper.py
@@ -1,28 +1,26 @@
 import os
 import json
-import time
-import uuid
 from os.path import join
 from pathlib import Path
 from meltano.core.utils import slugify
-
+from meltano.core.m5o.m5o_collection_parser import (
+    M5oCollectionParser,
+    M5oCollectionParserTypes,
+)
+from meltano.core.m5o.m5o_file_parser import MeltanoAnalysisFileParser
 from .sql_helper import SqlHelper
 
 
 class DashboardsHelper:
+    VERSION = "1.0.0"
+
     def __init__(self):
         self.meltano_model_path = join(os.getcwd(), "model")
-        self.dashboard_version = "0.1.0"
 
     def get_dashboards(self):
-        contents = []
-        dashboard_files = list(Path(self.meltano_model_path).glob("*.dashboard.m5o"))
-        for dashboard in dashboard_files:
-            file_name = dashboard.parts[-1]
-            file = Path(self.meltano_model_path).joinpath(file_name)
-            with file.open() as f:
-                contents.append(json.load(f))
-        return contents
+        path = Path(self.meltano_model_path)
+        dashboardsParser = M5oCollectionParser(path, M5oCollectionParserTypes.Dashboard)
+        return dashboardsParser.contents()
 
     def get_dashboard_reports_with_query_results(self, reports):
         sqlHelper = SqlHelper()
@@ -49,31 +47,33 @@ class DashboardsHelper:
         return target_dashboard[0]
 
     def save_dashboard(self, data):
-        data["id"] = uuid.uuid4().hex
-        data["version"] = self.dashboard_version
-        data["createdAt"] = time.time()
-        data["slug"] = slugify(data["name"])
-        data["reportIds"] = []
-        file_name = data["slug"] + ".dashboard.m5o"
+        slug = slugify(data["name"])
+        file_name = f"{slug}.dashboard.m5o"
         file_path = Path(self.meltano_model_path).joinpath(file_name)
+        data = MeltanoAnalysisFileParser.fill_base_m5o_dict(file_path, slug, data)
+        data["version"] = DashboardsHelper.VERSION
+        data["description"] = data["description"] or ""
+        data["reportIds"] = []
         with open(file_path, "w") as f:
             json.dump(data, f)
         return data
 
     def add_report_to_dashboard(self, data):
         dashboard = self.get_dashboard(data["dashboardId"])
-        dashboard["reportIds"].append(data["reportId"])
-        file_name = dashboard["slug"] + ".dashboard.m5o"
-        file_path = Path(self.meltano_model_path).joinpath(file_name)
-        with open(file_path, "w") as f:
-            json.dump(dashboard, f)
+        if data["reportId"] not in dashboard["reportIds"]:
+            dashboard["reportIds"].append(data["reportId"])
+            file_name = f"{dashboard['slug']}.dashboard.m5o"
+            file_path = Path(self.meltano_model_path).joinpath(file_name)
+            with open(file_path, "w") as f:
+                json.dump(dashboard, f)
         return dashboard
 
     def remove_report_from_dashboard(self, data):
         dashboard = self.get_dashboard(data["dashboardId"])
-        dashboard["reportIds"].remove(data["reportId"])
-        file_name = dashboard["slug"] + ".dashboard.m5o"
-        file_path = Path(self.meltano_model_path).joinpath(file_name)
-        with open(file_path, "w") as f:
-            json.dump(dashboard, f)
+        if data["reportId"] in dashboard["reportIds"]:
+            dashboard["reportIds"].remove(data["reportId"])
+            file_name = f"{dashboard['slug']}.dashboard.m5o"
+            file_path = Path(self.meltano_model_path).joinpath(file_name)
+            with open(file_path, "w") as f:
+                json.dump(dashboard, f)
         return dashboard