+++ src/meltano/api/controllers/settings_helper.py
@@ -1,5 +1,6 @@
 import os
 import json
+
 from os.path import join
 from pathlib import Path