+++ src/meltano/core/sql/analysis_helper.py
@@ -1,5 +1,6 @@
 from pypika import Table
 
+from .substitution import Substitution
 from .aggregate import Aggregate
 from .substitution import Substitution
 from .timeframe import TimeframePeriod