+++ CHANGELOG.md
@@ -9,14 +9,20 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ---
 
 ### New
+* Update installation and contributing docs
+* Meltano implement role-based access control - [!368](https://gitlab.com/meltano/meltano/merge_requests/368)
+* Add deep linking to dashboards
+* Add deep linking to reports
 
 ### Changes
 * Rename "Models" to "Topics"
 
 ### Fixes
+* Fixed a problem when environment variables where used as default values for the CLI - [!390](https://gitlab.com/meltano/meltano/merge_requests/390)
+* Fixed dashboards initial load issue due to legacy (and empty) `carbon.dashboard.m5o` file
 
 ### Breaks
-
+* New standardized approach for `.m5o` id generation (will need to remove any dashboard.m5o and report.m5o)
 
 ## 0.11.0 - (2019-02-19)
 ---
@@ -171,7 +177,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 * add default models for 'tap-carbon-intensity'.
 * Meltano Analyze is now part of the package.
 * removes database dependency from Meltano Analyze and uses .ma files
-* update the error message when using Meltano from outside a project - [238](https://gitlab.com/meltano/meltano/merge_requests/238)
+* update the error message when using Meltano from outside a project - [!238](https://gitlab.com/meltano/meltano/merge_requests/238)
 
 
 ## 0.3.0 - (2018-12-18)